+++ extensions/vscode/.gitignore
@@ -3,3 +3,4 @@
 node_modules/
 reference
 out/
+*.vsix
